requirements.txt
selenium==4.30.0
psycopg2-binary==2.9.9
paramiko==3.5.1
python-dotenv==1.0.1
sshtunnel==0.4.0
urllib3<2
